examples/flask_remote_app/requirements.txt
Flask>=0.11
python-etcd>=0.4
vyper
